refactor: 기존 lib 폴더 내부의 streak 관련 파일을 폴더로 묶음
