+++ apps/web/src/app/mypage/_components/voronoi-streak/voronoi-streak.tsx
@@ -9,9 +9,9 @@ import {
   VORONOI_COLOR_CONSTANT,
   VORONOI_NUMBER_CONSTANT,
 } from "../../_constants/voronoi-constant";
-import computeCells from "../../_lib/compute-cells";
-import generateRandomPoint from "../../_lib/generate-random-point";
-import lloydRelaxation from "../../_lib/lloyd-relaxation";
+import computeCells from "../../_lib/voronoi-streak/compute-cells";
+import generateRandomPoint from "../../_lib/voronoi-streak/generate-random-point";
+import lloydRelaxation from "../../_lib/voronoi-streak/lloyd-relaxation";
 import { YearlyAnswerSubmissions } from "../../_types/streak";
 
 interface VoronoiStreakProps {
